+++ src/features/dashboard/views/CampaignResultDashboardView.vue
